backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
python-multipart
jsonschema
python-dotenv
httpx
pymupdf
python-pptx
